Update InfoPage to MVVM pattern. Relocated Localization to NureTimetable.Core project
