+++ NureTimetable.Core/Localization/LN.resx
