+++ NureTimetable.Core/Localization/LN.uk.resx
