+++ NureTimetable.sln
@@ -9,7 +9,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NureTimetable", "NureTimeta
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NureTimetable.Services", "NureTimetable.Services\NureTimetable.Services.csproj", "{7C19C2E3-144F-4909-A3BA-282FED3EF282}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NureTimetable.UI", "NureTimetable.UI\NureTimetable.UI.csproj", "{D06BF130-7991-445B-9B32-0BE8299EEE12}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NureTimetable.UI", "NureTimetable.UI\NureTimetable.UI.csproj", "{D06BF130-7991-445B-9B32-0BE8299EEE12}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NureTimetable.Core", "NureTimetable.Core\NureTimetable.Core.csproj", "{7212568B-A874-4DDF-AA58-67A8D57D0F81}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -37,6 +39,10 @@ Global
 		{D06BF130-7991-445B-9B32-0BE8299EEE12}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D06BF130-7991-445B-9B32-0BE8299EEE12}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D06BF130-7991-445B-9B32-0BE8299EEE12}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7212568B-A874-4DDF-AA58-67A8D57D0F81}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7212568B-A874-4DDF-AA58-67A8D57D0F81}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7212568B-A874-4DDF-AA58-67A8D57D0F81}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7212568B-A874-4DDF-AA58-67A8D57D0F81}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
