+++ NureTimetable/NureTimetable/NureTimetable.csproj
@@ -22,11 +22,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="Licalization\LN.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>LN.resx</DependentUpon>
-    </Compile>
     <ProjectReference Include="..\..\NureTimetable.Services\NureTimetable.Services.csproj" />
     <ProjectReference Include="..\..\NureTimetable.UI\NureTimetable.UI.csproj" />
   </ItemGroup>
@@ -38,10 +33,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <EmbeddedResource Update="Licalization\LN.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>LN.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
     <EmbeddedResource Update="Views\Info\DonatePage.xaml">
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
